Extract `database` and `settings` modules
